+++ axle-core/src/main/scala/axle/visualize/BarChartGroupedView.scala
@@ -36,8 +36,8 @@ case class BarChartGroupedView[G, S, Y, D, H](
   implicit val ddls = axle.algebra.LengthSpace.doubleDoubleLengthSpace
 
   val scaledArea = ScaledArea2D(
-    border, (if (drawKey) width - (keyWidth + keyLeftPadding) else width) - border,
-    border, height - (if (labelAngle.isDefined) border else 5),
+    border.toDouble, (if (drawKey) width - (keyWidth + keyLeftPadding) else width) - border.toDouble,
+    border.toDouble, height - (if (labelAngle.isDefined) border.toDouble else 5d),
     minX, maxX,
     minY, maxY)
 
@@ -48,13 +48,13 @@ case class BarChartGroupedView[G, S, Y, D, H](
     scaledArea,
     groups.toStream.zipWithIndex.map({ case (g, i) => (padding + (i + 0.5) * widthPerGroup, string(g)) }).toList,
     normalFontName,
-    normalFontSize,
+    normalFontSize.toDouble,
     bold = true,
     drawLines = false,
     labelAngle,
     black)
 
-  val yTics = YTics(scaledArea, Tics[Y].tics(minY, maxY), normalFontName, normalFontSize, true, black)
+  val yTics = YTics(scaledArea, Tics[Y].tics(minY, maxY), normalFontName, normalFontSize.toDouble, true, black)
 
   val barSliceWidth = (widthPerGroup - (whiteSpace / 2d)) / slices.size.toDouble
