+++ axle-core/src/main/scala/axle/visualize/PlotView.scala
@@ -34,8 +34,8 @@ case class PlotView[S, X, Y, D](
 
   val vLine = VerticalLine(scaledArea, yAxis.getOrElse(minX), black)
   val hLine = HorizontalLine(scaledArea, xAxis.getOrElse(minY), black)
-  val xTics = XTics(scaledArea, xts.tics(minX, maxX), fontName, fontSize, bold = true, drawLines = true, Some(0d *: angleDouble.degree), black)
-  val yTics = YTics(scaledArea, yts.tics(minY, maxY), fontName, fontSize, true, black)
+  val xTics = XTics(scaledArea, xts.tics(minX, maxX), fontName, fontSize.toDouble, bold = true, drawLines = true, Some(0d *: angleDouble.degree), black)
+  val yTics = YTics(scaledArea, yts.tics(minY, maxY), fontName, fontSize.toDouble, true, black)
 
   val dataLines = DataLines(scaledArea, data, plotDataView.xsOf, plotDataView.valueOf, colorOf, pointDiameter, connect)
